Predictive Maintenance/data-generator/requirements.txt
pandas
numpy
paho-mqtt
